+++ Vribbels/ui/tabs/__init__.py
@@ -8,5 +8,6 @@ from .optimizer_tab import OptimizerTab
 from .heroes_tab import HeroesTab
 from .scoring_tab import ScoringTab
 from .about_tab import AboutTab
+from .rescue_tab import RescueTab
 
-__all__ = ['MaterialsTab', 'SetupTab', 'CaptureTab', 'InventoryTab', 'OptimizerTab', 'HeroesTab', 'ScoringTab', 'AboutTab']
+__all__ = ['MaterialsTab', 'SetupTab', 'CaptureTab', 'InventoryTab', 'OptimizerTab', 'HeroesTab', 'ScoringTab', 'AboutTab', 'RescueTab']